+++ tests/date.test.ts
@@ -1,13 +1,20 @@
-import { assertStrictEquals as equals } from "../deps/assert.ts";
+import { assert, assertStrictEquals as equals } from "../deps/assert.ts";
 import lume from "../mod.ts";
 import date from "../plugins/date.ts";
+import { modulePath } from "../deps/date.ts";
+import { build, getDepVersion, getSite } from "./utils.ts";
 import gl from "https://deno.land/x/date_fns@v2.22.1/locale/gl/index.js";
 import pt from "https://deno.land/x/date_fns@v2.22.1/locale/pt/index.js";
 
 Deno.env.set("TZ", "Z");
 const date0 = new Date(Date.UTC(1970, 0, 1));
 
-Deno.test("date plugin", () => {
+Deno.test("date_fn version", async () => {
+  const version = await getDepVersion("date.ts", "date_fns");
+  equals(`https://deno.land/x/date_fns@${version}`, modulePath);
+});
+
+Deno.test("date plugin", async () => {
   const site = lume();
   site.use(date({
     formats: {
@@ -16,6 +23,11 @@ Deno.test("date plugin", () => {
   }));
 
   const { helpers } = site.engines;
+
+  assert(!helpers.has("date"));
+  await site.dispatchEvent({ type: "beforeBuild" });
+  assert(helpers.has("date"));
+
   const [format] = helpers.get("date")!;
 
   equals(format(date0, "yyyy-MM-dd"), "1970-01-01");
@@ -33,11 +45,13 @@ Deno.test("date plugin", () => {
   equals(format(date0, "CUSTOM"), "1970_01");
 });
 
-Deno.test("date plugin with custom locale", () => {
+Deno.test("date plugin with custom locale", async () => {
   const site = lume();
   site.use(date({
     locales: { gl, pt },
   }));
+
+  await site.dispatchEvent({ type: "beforeBuild" });
 
   const { helpers } = site.engines;
   const [format] = helpers.get("date")!;
@@ -54,15 +68,35 @@ Deno.test("date plugin with custom locale", () => {
   );
 });
 
-Deno.test("date plugin with custom name", () => {
+Deno.test("date plugin with custom name", async () => {
   const site = lume();
   site.use(date({
     name: "dateify",
     locales: { gl, pt },
   }));
 
+  await site.dispatchEvent({ type: "beforeBuild" });
+
   const { helpers } = site.engines;
   const [format] = helpers.get("dateify")!;
 
   equals(format(date0, "HUMAN_DATE"), "1 de xaneiro 1970");
+});
+
+Deno.test("date plugin load locales automatically", async () => {
+  const site = getSite({
+    src: "simple",
+  });
+
+  site.use(date({
+    locales: ["gl", "pt"],
+  }));
+
+  await build(site);
+
+  const { helpers } = site.engines;
+
+  const [format] = helpers.get("date")!;
+  equals(format(date0, "HUMAN_DATE"), "1 de xaneiro 1970");
+  equals(format(date0, "HUMAN_DATE", "pt"), "1 de janeiro de 1970");
 });