+++ tests/utils.test.ts
@@ -1,8 +1,6 @@
 import { assertEquals as equals } from "../deps/assert.ts";
 import { getImportMap, isPlainObject, merge, sha1 } from "../core/utils.ts";
-import { getDepVersion, getPath } from "./utils.ts";
-
-const datefn_version = await getDepVersion("date.ts", "date_fns");
+import { getPath } from "./utils.ts";
 
 Deno.test("merge options", () => {
   interface Options {
@@ -88,8 +86,6 @@ Deno.test("import map", async () => {
       "lume": new URL("../mod.ts", import.meta.url).href,
       "lume/": new URL("../", import.meta.url).href,
       "https://deno.land/x/lume/": new URL("../", import.meta.url).href,
-      "lume/plugins/date/locale":
-        `https://deno.land/x/date_fns@${datefn_version}/locale`,
     },
   });
 });
@@ -102,8 +98,6 @@ Deno.test("merge import map", async () => {
       "lume": new URL("../mod.ts", import.meta.url).href,
       "lume/": new URL("../", import.meta.url).href,
       "https://deno.land/x/lume/": new URL("../", import.meta.url).href,
-      "lume/plugins/date/locale":
-        `https://deno.land/x/date_fns@${datefn_version}/locale`,
       "std/": "https://deno.land/std@0.121.0/",
       "/": "./",
     },