+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+### Added
+- `site.use()` accepts Promises, resolved in the `beforeBuild` event.
+- `date` plugin accepts an array of locales (ex: `["en", "gl", "pt", "es"]`),
+  that are loaded automatically.
+
+### Removed
+- `lume/plugins/date/locale` import from the import_map.
+
 ### Fixed
 - Throw an exception on missing configuration for Netlify CMS in the `netlify_cms` plugin.
 - Fixed `expires` middleware for `0` values.